Code Improve - move `get_parameters_with_values` and `remove_parameter` to `projects.utils` due circular import when calling from `projects.models.operator` - change `update_status` to be an Operator hybrid property - use `assertCountEqual` to validate list
