+++ projects/controllers/operators.py
@@ -5,14 +5,11 @@ from datetime import datetime
 from sqlalchemy.exc import InvalidRequestError, ProgrammingError
 from werkzeug.exceptions import BadRequest, NotFound
 
-from projects.controllers.tasks.parameters import list_parameters, \
-    get_parameters_with_values, remove_parameter
 from projects.controllers.utils import raise_if_task_does_not_exist, \
     raise_if_project_does_not_exist, raise_if_experiment_does_not_exist, \
     raise_if_operator_does_not_exist, uuid_alpha
 from projects.database import db_session
-from projects.kfp.runs import get_container_status
-from projects.models import Operator, Task
+from projects.models import Operator
 
 
 PARAMETERS_EXCEPTION_MSG = "The specified parameters are not valid"
@@ -47,7 +44,6 @@ def list_operators(project_id, experiment_id):
 
     response = []
     for operator in operators:
-        update_status(operator)
         response.append(operator.as_dict())
 
     return response
@@ -125,8 +121,6 @@ def create_operator(project_id, experiment_id=None, deployment_id=None,
     db_session.add(operator)
     db_session.commit()
 
-    update_status(operator)
-
     return operator.as_dict()
 
 
@@ -176,8 +170,6 @@ def update_operator(project_id, experiment_id, operator_id, **kwargs):
         db_session.commit()
     except (InvalidRequestError, ProgrammingError) as e:
         raise BadRequest(str(e))
-
-    update_status(operator)
 
     return operator.as_dict()
 
@@ -357,34 +349,3 @@ def has_cycles_util(operator_id, visited, recursion_stack, new_dependencies, new
 
     recursion_stack[operator_id] = False
     return False
-
-
-def update_status(operator):
-    """
-    Update operator status.
-
-    Parameters
-    ----------
-    operator : projects.models.operator.Operator
-    """
-    # get total operator parameters with value
-    parameters = get_parameters_with_values(operator.parameters)
-    task = Task.query.get(operator.task_id)
-    status = get_container_status(operator.experiment_id, operator.uuid)
-
-    if status:
-        operator.status = status
-    elif "DATASETS" in task.tags:
-        if parameters:
-            operator.status = "Setted up"
-        else:
-            operator.status = "Unset"
-    else:
-        # get task parameters and remove dataset parameter
-        task_parameters = list_parameters(operator.task_id)
-        task_parameters = remove_parameter(task_parameters, "dataset")
-
-        if len(parameters) == len(task_parameters):
-            operator.status = "Setted up"
-        else:
-            operator.status = "Unset"
